Split `emotion`
