+++ packages/next-swc/crates/core/Cargo.toml
@@ -8,18 +8,16 @@ publish = false
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-base64 = "0.13"
-byteorder = "1"
 chrono = "0.4"
 easy-error = "1.0.0"
 either = "1"
 fxhash = "0.2.1"
 once_cell = "1.8.0"
 pathdiff = "0.2.0"
-radix_fmt = "1"
 regex = "1.5"
 serde = "1"
 serde_json = "1"
+emotion = {path="../emotion"}
 styled_components = {path="../styled_components"}
 modularize_imports = {path="../modularize_imports"}
 swc = "0.161.1"
