+++ packages/next-swc/crates/emotion/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+authors = ["next.js developers"]
+edition = "2018"
+license = "Apache-2.0"
+name = "emotion"
+repository = "https://github.com/vercel/next.js.git"
+version = "0.1.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+base64 = "0.13"
+byteorder = "1"
+fxhash = "0.2.1"
+once_cell = "1.8.0"
+radix_fmt = "1"
+regex = "1.5"
+serde = "1"
+sourcemap = "6.0.1"
+swc_atoms = "0.2.11"
+swc_common = {version = "0.17.18", features = ["concurrent", "sourcemap"]}
+swc_ecmascript = {version = "0.140.0", features = ["codegen", "utils", "visit"]}
+
+[dev-dependencies]
+swc_ecma_transforms_testing = "0.75.0"
+testing = "0.19.1"
