+++ packages/next-swc/crates/emotion/tests/fixture/compress/input.tsx
