+++ packages/next-swc/crates/emotion/tests/fixture/compress/output.ts
