+++ packages/next-swc/crates/emotion/tests/fixture/namespace-import/input.tsx
