+++ packages/next-swc/crates/emotion/tests/fixture/namespace-import/output.ts
